api: move RAFS configuration struct into api
Move RAFS configuration struct into api.
