+++ blobfs/Cargo.toml
@@ -17,6 +17,7 @@ serde_json = "1.0.53"
 vm-memory = { version = "0.9" }
 
 nydus-error = { version = "0.2", path = "../error" }
+nydus-api = { version = "0.1", path = "../api" }
 nydus-rafs = { version = "0.1", path = "../rafs" }
 nydus-storage = { version = "0.5", path = "../storage", features = ["backend-localfs"] }
